Clean Model

+++ src/Model/ModelOptimLikelihood.cpp
@@ -54,7 +54,7 @@ bool ModelOptimLikelihood::_checkConsistency()
   const Model* model = _modelPart._model;
   const Db* db       = _dbPart._db;
 
-  if (model->getNDim() != db->getNLoc(ELoc::X))
+  if ((int)model->getNDim() != db->getNLoc(ELoc::X))
   {
     messerr("'_model'(%d) and 'db'(%d) should have same Space Dimension",
             model->getNDim(), db->getNLoc(ELoc::X));
